add dmrpp docs to top level docs
